08.16 remove urls & models
