+++ core/accounts/api/v1/urls/accounts.py
@@ -1,10 +1,8 @@
 from django.urls import path
-from . import views
+from .. import views
 from rest_framework_simplejwt.views import (TokenRefreshView,TokenVerifyView)
 # from rest_framework.authtoken.views import ObtainAuthToken
 
-
-app_name = 'api-v1'
 
 urlpatterns = [
     # registration
@@ -24,14 +22,6 @@ urlpatterns = [
     # login jwt
     path('jwt/create/', views.CustomTokenObtainPairView.as_view(), name='jwt-create'),
     path('jwt/refresh/', TokenRefreshView.as_view(), name='jwt-refresh'),
-    path('jwt/verify/', TokenVerifyView.as_view(), name='jwt-verify'),
-
-    
-    # profile
-    path('profile/', views.ProfileApiView.as_view(), name='profile')
-    
+    path('jwt/verify/', TokenVerifyView.as_view(), name='jwt-verify'),    
     
 ]
-
-
-
